requirements.txt
pyautogui==0.9.54
keyboard==0.13.5
pywin32==306
configparser==6.0.0
pynput==1.7.6
opencv-python==4.8.0.76
Pillow==10.0.0
PyScreeze==0.1.29
python-dotenv==1.0.1
prettytable==3.14.0 # for add_rows() support
PTable==0.9.2
matplotlib==3.8.0
playsound==1.2.2 # downgrade to avoid bug
discord.py==2.4.0

# reference: https://stackoverflow.com/questions/68704443/python-playsound-error-261-for-command-the-driver-cannot-recognize-the-specifie